Fix signedness warning.
git-svn-id: svn+ssh://svn.dolda2000.com/srv/svn/repos/src/doldaconnect@707 959494ce-11ee-0310-bf91-de5d638817bd

+++ include/doldaconnect/uilib.h
@@ -15,7 +15,7 @@ struct dc_response
 	int argc;
 	wchar_t **argv;
     } *rlines;
-    int linessize;
+    size_t linessize;
     int numlines;
     int curline;
 };
